# frontend/src/types/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RideCompass API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/weather": {
      "get": {
        "operationId": "get_weather_api_weather_get",
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "title": "Latitude"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "title": "Longitude"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherConditions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/weather/warnings": {
      "get": {
        "operationId": "get_weather_warnings_api_weather_warnings_get",
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "title": "Latitude"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "title": "Longitude"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherWarnings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/weather/wbgt": {
      "get": {
        "operationId": "get_wbgt_api_weather_wbgt_get",
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "title": "Latitude"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "title": "Longitude"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WbgtStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/weather/flood-forecast": {
      "get": {
        "operationId": "get_flood_forecast_api_weather_flood_forecast_get",
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "title": "Latitude"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "title": "Longitude"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloodForecasts"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/weather/amedas": {
      "get": {
        "operationId": "get_amedas_api_weather_amedas_get",
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "title": "Latitude"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "title": "Longitude"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AmedasObservation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/weather/wind-grid-detail": {
      "get": {
        "operationId": "get_wind_grid_detail_api_weather_wind_grid_detail_get",
        "parameters": [
          {
            "name": "min_lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "title": "Min Lon"
            }
          },
          {
            "name": "min_lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "title": "Min Lat"
            }
          },
          {
            "name": "max_lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "title": "Max Lon"
            }
          },
          {
            "name": "max_lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "title": "Max Lat"
            }
          },
          {
            "name": "spacing_deg",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.02,
              "title": "Spacing Deg"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WindGridResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActiveFloodForecast": {
        "properties": {
          "river_code": {
            "type": "string",
            "title": "River Code"
          },
          "badge_level": {
            "type": "string",
            "enum": [
              "advisory",
              "warning",
              "severe_warning",
              "emergency_warning"
            ],
            "title": "Badge Level"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "condition": {
            "type": "string",
            "title": "Condition"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "river_code",
          "badge_level",
          "label",
          "condition"
        ],
        "title": "ActiveFloodForecast"
      },
      "ActiveWarning": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "level": {
            "type": "string",
            "enum": [
              "advisory",
              "warning",
              "severe_warning",
              "emergency_warning"
            ],
            "title": "Level"
          },
          "additions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Additions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "name",
          "level",
          "additions"
        ],
        "title": "ActiveWarning"
      },
      "AmedasObservation": {
        "properties": {
          "station_name": {
            "type": "string",
            "title": "Station Name"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature C"
          },
          "apparent_temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apparent Temperature C"
          },
          "wind_speed_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Speed Ms"
          },
          "wind_direction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WindDirection"
              },
              {
                "type": "null"
              }
            ]
          },
          "precipitation_10min_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precipitation 10Min Mm"
          },
          "twilight": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Twilight"
              },
              {
                "type": "null"
              }
            ]
          },
          "weather_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weather Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "station_name",
          "observed_at",
          "temperature_c",
          "apparent_temperature_c",
          "wind_speed_ms",
          "wind_direction",
          "precipitation_10min_mm",
          "twilight",
          "weather_code"
        ],
        "title": "AmedasObservation"
      },
...
      "FloodForecasts": {
        "properties": {
          "forecasts": {
            "items": {
              "$ref": "#/components/schemas/ActiveFloodForecast"
            },
            "type": "array",
            "title": "Forecasts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "forecasts"
        ],
        "title": "FloodForecasts"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TemperatureRange": {
        "properties": {
          "min_c": {
            "type": "number",
            "title": "Min C"
          },
          "max_c": {
            "type": "number",
            "title": "Max C"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "min_c",
          "max_c"
        ],
        "title": "TemperatureRange"
      },
...
      "Twilight": {
        "properties": {
          "sunrise": {
            "type": "string",
            "title": "Sunrise"
          },
          "sunset": {
            "type": "string",
            "title": "Sunset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sunrise",
          "sunset"
        ],
        "title": "Twilight"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WbgtReading": {
        "properties": {
          "level": {
            "type": "string",
            "enum": [
              "advisory",
              "warning",
              "severe_warning",
              "emergency_warning"
            ],
            "title": "Level"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "number",
            "title": "Value"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "level",
          "label",
          "value"
        ],
        "title": "WbgtReading"
      },
      "WbgtStatus": {
        "properties": {
          "reading": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WbgtReading"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reading"
        ],
        "title": "WbgtStatus"
      },
      "WeatherConditions": {
        "properties": {
          "precipitation_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precipitation Mm"
          },
          "twilight": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Twilight"
              },
              {
                "type": "null"
              }
            ]
          },
          "precipitation_max_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precipitation Max Mm"
          },
          "wind_speed_max_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Speed Max Ms"
          },
          "temperature_range": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TemperatureRange"
              },
              {
                "type": "null"
              }
            ]
          },
          "today_periods": {
            "items": {
              "$ref": "#/components/schemas/WeatherPeriodOutlook"
            },
            "type": "array",
            "title": "Today Periods"
          },
          "today_period_interval_hours": {
            "type": "integer",
            "title": "Today Period Interval Hours"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "precipitation_mm",
          "twilight",
          "precipitation_max_mm",
          "wind_speed_max_ms",
          "temperature_range",
          "today_periods",
          "today_period_interval_hours"
        ],
        "title": "WeatherConditions"
      },
      "WeatherPeriodOutlook": {
        "properties": {
          "period": {
            "type": "string",
            "title": "Period"
          },
          "temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature C"
          },
          "precipitation_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precipitation Mm"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "period",
          "temperature_c",
          "precipitation_mm"
        ],
        "title": "WeatherPeriodOutlook"
      },
      "WeatherWarnings": {
        "properties": {
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ActiveWarning"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "warnings"
        ],
        "title": "WeatherWarnings"
      },
      "WindDirection": {
        "properties": {
          "deg": {
            "type": "number",
            "title": "Deg"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "deg",
          "label"
        ],
        "title": "WindDirection"
      },
      "WindGridPoint": {
        "properties": {
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "wind_speed_ms": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "title": "Wind Speed Ms"
          },
          "wind_direction_deg": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "title": "Wind Direction Deg"
          },
          "precipitation_mm": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "title": "Precipitation Mm"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "wind_speed_ms",
          "wind_direction_deg",
          "precipitation_mm"
        ],
        "title": "WindGridPoint"
      },
      "WindGridResponse": {
        "properties": {
          "times": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Times"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/WindGridPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "times",
          "points"
        ],
        "title": "WindGridResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}